Add "actions" system
This allows pages representing actions (e.g. new, edit, etc.) to be indicated by GET attributes (or the _action POST attribute) instead of as subpages in the URL. As subpages, they could have subpages of their own which wouldn't ever make sense, but as GET/POST attributes, this is no longer the case.